% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top3_high.R
\name{top3_high}
\alias{top3_high}
\title{Calculate greatest top 3}
\usage{
top3_high(arr)
}
\arguments{
\item{arr}{the target array in which we need to find the greatest 3 number}
}
\value{
high3
}
\description{
Calculate greatest top 3 numbers in a given list
}
\details{
This function calculates the greatest top 3 numbers in a list by bubble sort
}
\examples{
top3_high(1:5)

}
