% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top3_low.R
\name{top3_low}
\alias{top3_low}
\title{Calculate smallest top 3}
\usage{
top3_low(arr)
}
\arguments{
\item{arr}{the target array in which we need to find the smallest 3 number}
}
\value{
low3
}
\description{
Calculate smallest top 3 numbers in a given list
}
\details{
This function calculates the smallest top 3 numbers in a list by bubble sort
}
\examples{
top3_low(c(1,3,5,7,9))

}
